refactored the db orm with SQLAlchemy
This will be better for the long term, and will greatly simplify the
complexity of the data structure, particularly due to the large number
of columns on alot of the planned schemas.

This is highly scaleable and very minimal, also changed a few of the API
endpoints to fit with the new db structure.

One thing to keep in mind is that we are not able to natively serialize
DB models into JSON format, so I made a serializable class that we need
to mixin to allow for quick and easy json serialization.