fixed an issue with reset password endpoint
it was previously taking the arguments as query parameters, where as
that is very unsafe.
It now takes the arguments in a post request.

Made a new object model to handle this specific request to accomodate
this fix.

also fixed an issue with the JWT encoder that requires an additional
parameter.